backend/requirements/base.txt
# Core dependencies
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
python-dotenv>=1.0,<2.0
